refactor(rewrite): rename types.py to common.py in reorder_static_placement
