+++ python/bloqade/lanes/rewrite/reorder_static_placement/common.py
